modification call db for getProfile api

+++ src/backend/typescript/services/SuscriberService.ts
@@ -1,4 +1,4 @@
-import { PrismaClient, type User } from "@prisma/client";
+import { PrismaClient, type User, type Friendship, type Match  } from "@prisma/client";
 import { type UpdateData } from "../types/suscriber.types.js";
 import { PasswordHasher } from "../utils/PasswordHasher.js";
 import { FileService } from "./FileService.js";
@@ -7,6 +7,8 @@ import type { SuscriberProfile } from "../types/suscriber.types.js";
 import type { GameStats } from "../types/match.types.js";
 import { SuscriberException, SuscriberError } from "../error_handlers/Suscriber.error.js";
 import path from "path";
+import { SocketEventController } from "../controllers/SocketEventController.js";
+import type { Friend, MatchSummary } from "../types/suscriber.types.js";
 
 export class SuscriberService {
     constructor(
@@ -15,17 +17,32 @@ export class SuscriberService {
         private fileService: FileService
     ) {}
     private api_url = process.env.API_URL || 'https://localhost:8080/api';
-    private default_avatar_url = this.api_url + '/static/public/avatarDefault.webp';
-
+    private default_avatar_filename = 'avatarDefault.webp';
+    private default_avatar_url = this.api_url + '/static/public/' + this.default_avatar_filename;
     // ----------------------------------------------------------------------------- //
     async getProfile(id: number): Promise<SuscriberProfile> {
+        // we are limited to 10 matches won and 10 matches lost to calculate stats and get last matches
+        const takeLimit = 10; 
+
+        // we count matches won and lost for stats calculation
         const user = await this.prisma.user.findUnique({
             where: { id: Number(id) },
             include: {
-                matchesWons: true,
-                matchesLoses: true,
-                sentRequests: true,
-                receivedRequests: true,
+                _count: {
+                    select: { matchesWons: true, matchesLoses: true }
+                },
+                matchesWons: { 
+                    include: { loser: true },
+                    orderBy: { createdAt: 'desc' },
+                    take: takeLimit
+                },
+                matchesLoses: { 
+                    include: { winner: true },
+                    orderBy: { createdAt: 'desc' },
+                    take: takeLimit
+                },
+                sentRequests: { include: { receiver: true } }, 
+                receivedRequests: { include: { requester: true } }, 
             },
         });
 
@@ -33,27 +50,14 @@ export class SuscriberService {
             throw new SuscriberException(SuscriberError.USER_NOT_FOUND, SuscriberError.USER_NOT_FOUND);
         }
 
-		const stats: GameStats = this.calculateStats(user);
+		const stats: GameStats = this.calculateStats(user._count.matchesWons, user._count.matchesLoses);
 
 		// Get last 4 matches
-        const allMatches = [
-            ...(user.matchesWons || []),
-            ...(user.matchesLoses || []),
-        ]
-		.sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime())
-        .slice(0, 4);
+        const allMatches = this.getLastMatches(user.matchesWons, user.matchesLoses, 4);
 
 		// Sorted friendships
-        const sortedFriendships = [
-            ...(user.sentRequests || []),
-            ...(user.receivedRequests || []),
-        ]
-		.sort((a, b) => {
-			if (a.status === "PENDING" && b.status !== "PENDING") return -1;
-			if (a.status !== "PENDING" && b.status === "PENDING") return 1;
-			return 0;
-		})
-		.slice(0, 4);
+        const sortedFriendships = this.getSortedFriendlist(user.sentRequests, user.receivedRequests, 4, user.id);
+
 
         return {
             id: user.id.toString(),
@@ -116,14 +120,12 @@ export class SuscriberService {
             where: { id: Number(id) },
             data: {
                 username: data.username ?? user.username,
-                avatar: data.avatar ?? user.avatar
             },
         });
 
         return sanitizeUser(updatedUser);
     }
 
-	// a verifier
     // ----------------------------------------------------------------------------- //
     async updateAvatar(buffer: Buffer, userId: number): Promise<SanitizedUser> {
         const user = await this.getById(userId);
@@ -136,16 +138,16 @@ export class SuscriberService {
             throw new SuscriberException(SuscriberError.UPLOAD_ERROR, SuscriberError.UPLOAD_ERROR);
         
         const oldAvatarUrl = user.avatar;
-        const newAvatarPath = path.join('/static/avatars/', path.basename(avatarFileName));
-        const newAvatarUrl = this.api_url + newAvatarPath;
+        const newAvatarUrl = `${this.api_url}/static/avatars/${path.basename(avatarFileName)}`;
         try {
             const updatedUser = await this.prisma.user.update({
                 where: { id: userId },
                 data: { avatar: newAvatarUrl },
             });
             
-            await this.fileService.deleteAvatar(oldAvatarUrl);
-
+            if (!oldAvatarUrl.includes(this.default_avatar_filename)) {
+                await this.fileService.deleteAvatar(oldAvatarUrl);
+            }
             return sanitizeUser(updatedUser);
 
         } catch (error) {
@@ -170,7 +172,9 @@ export class SuscriberService {
             data: { avatar: this.default_avatar_url },
         });
 
-        await this.fileService.deleteAvatar(oldAvatarUrl);
+        if (!oldAvatarUrl.includes(this.default_avatar_filename)) {
+            await this.fileService.deleteAvatar(oldAvatarUrl);
+        }
 
         return sanitizeUser(updatedUser);
     }
@@ -191,7 +195,7 @@ export class SuscriberService {
 
     // ----------------------------------------------------------------------------- //
     private async getById(id: number): Promise<User | null> {
-        return await this.prisma.user.findFirst({ where: { id } });
+        return await this.prisma.user.findUnique({ where: { id } });
     }
 
     // ----------------------------------------------------------------------------- //
@@ -200,18 +204,103 @@ export class SuscriberService {
     }
 	
 	// ----------------------------------------------------------------------------- //
-	private calculateStats(user: any): GameStats {
-		const matchesWon = user.matchesWons?.length || 0;
-		const matchesLost = user.matchesLoses?.length || 0;
-		const totalMatches = matchesWon + matchesLost;
-		const ratio = totalMatches > 0 ? (matchesWon / totalMatches * 100).toFixed(2) : "0.00";
+	private calculateStats(matchesWons: number, matchesLoses: number): GameStats {
+		const totalMatches = matchesWons + matchesLoses;
+		const ratio = totalMatches > 0 ? (matchesWons / totalMatches * 100).toFixed(2) : "0.00";
 
 		return {
-			wins: matchesWon,
-			losses: matchesLost,
+			wins: matchesWons,
+			losses: matchesLoses,
 			total: totalMatches,
 			winRate: parseFloat(ratio),
 		};
 	}
-}
 
+	// ----------------------------------------------------------------------------- //
+    private getLastMatches(matchesWons: any[], matchesLoses: any[], limit: number): MatchSummary[] {
+        /**
+         * On combine les deux listes de matchs (gagnés et perdus),
+         * on les trie par date de création décroissante,
+         * puis on prend les 'limit' premiers.
+         * on retourne une structure combinée { opponent, match }
+         * 
+         * on prend en compte que l'utilisateur peut ne pas avoir de match
+         * ou bien aucun match gagne
+         * ou bien aucun perdu
+         * 
+         */
+        const allMatches = [
+            ...(matchesWons || []),
+            ...(matchesLoses || []),
+        ].sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime())
+        .slice(0, limit);
+
+        // Transformation des données pour le front
+        return allMatches.map(match => {
+            // si l'utilisateur est supprimé (SetNull), opponentObj peut être null.
+            const opponentObj = match.loser || match.winner;
+
+            return {
+                opponent: opponentObj ? {
+                    id: opponentObj.id.toString(),
+                    username: opponentObj.username,
+                    avatar: opponentObj.avatar
+                } : null, // Cas où l'adversaire a supprimé son compte
+                match: match as Match,
+            } as MatchSummary;
+        });
+    }
+
+    // ----------------------------------------------------------------------------- //
+    private getSortedFriendlist(requester: any[], receiver: any[], limit: number, userId: number): Friend[] {
+        /**
+         * dans cette fonction on doit trier les amis selon ces critères :
+         * 1. Les amis avec le statut 'ACCEPTED' et qui sont en ligne (connectés via websocket) doivent apparaître en premier.
+         * 2. Ensuite, les demandes d'amitié en attente avec le statut 'PENDING'.
+         * 3. Enfin, les amis avec le statut 'ACCEPTED' mais qui sont hors ligne.
+         * ...
+         */
+        const allFriendships = [
+            ...(requester || []),
+            ...(receiver || []),
+        ];
+
+        // 1. Tri
+        allFriendships.sort((a, b) => {
+            const getFriendId = (r: Friendship) => r.requesterId === userId ? r.receiverId : r.requesterId;
+            
+            const isOnlineA = SocketEventController.isUserOnline(getFriendId(a));
+            const isOnlineB = SocketEventController.isUserOnline(getFriendId(b));
+
+            /* Priority Score:
+               2 -> Ami (ACCEPTED) et Connecté
+               1 -> Demande en attente (PENDING)
+               0 -> Ami déconnecté ou autre
+            */
+            const getScore = (r: any, isOnline: boolean) => {
+                if (r.status === 'ACCEPTED' && isOnline) return 2;
+                if (r.status === 'PENDING') return 1;
+                return 0;
+            };
+
+            return getScore(b, isOnlineB) - getScore(a, isOnlineA);
+        });
+
+        // 2. Limite
+        const sliced = allFriendships.slice(0, limit);
+
+        // 3. Transformation / Nettoyage
+        return sliced.map(relation => {
+            // On détermine qui est l'ami (l'autre personne dans la relation)
+            const friend = relation.requesterId === userId ? relation.receiver : relation.requester;
+            
+            return {
+                id: friend.id.toString(),
+                username: friend.username,
+                avatar: friend.avatar,
+                isOnline: SocketEventController.isUserOnline(friend.id),
+                status: relation.status,
+            };
+        });
+    }
+}
